feature: add --no-delete option to chkbit update command.
This feature adds support for not deleting missing files or
directories from the index (default is to remove them). This
is for cases where the mount point might to temprary offline
and you don't want to reindex them again when it is online
and are using SkipExisting features, thus saves a lot of time
if the repo exceeds a couple of terrabyte for those specific
directores/files.